# perf(examples-index): ditch Pageres and use Puppeteer directly