--- a/src/generate-examples-index/content-builder/index.ts
+++ b/src/generate-examples-index/content-builder/index.ts
@@ -2,3 +2,4 @@ export * from "./builder";
 export { generatePlaygroundData } from "./playground";
 export { htmlRenderer } from "./html";
 export { mdRenderer } from "./md";
+export { formatMs, formatStartStopMs, reduceReports } from "./common";