--- a/tsconfig.json
+++ b/tsconfig.json
@@ -3,7 +3,7 @@
     "baseUrl": ".",
     "declaration": false,
     "esModuleInterop": true,
-    "lib": ["es2020"],
+    "lib": ["dom", "es2020"],
     "module": "es2015",
     "moduleResolution": "node",
     "outDir": "js/src",