Currently generating screenshots for examples index in Vis Network takes over 10 minutes, usually around 13 to 14 minutes; with this it gets bellow 2 minutes. Atop of this since I use single browser for all of them, there isn't that much overhead for taking multiple screenshots in parallel (opening new tab vs starting new browser). With Pageres the performance didn't increase and even deteriorated above 6 parallel runs; using Puppeteer directly, I'm able to run all of them in parallel and get all the screenshots in around 30 seconds.

It's also now possible for the screenshots to be taken when the page says it's ready (calling `window.takeScreenshot()` from the example). This feature was the main reason why I did this, but apparently most of the slowdown was due to the way Pageres works, the benefit of this would likely be negligible and not worth it at the moment. I kept it in the code as it may come in handy in the future.

The times taken are now reported to pin point what takes the most time, though this mostly shows that it doesn't matter that much what is on the page, they all take more or less the same amount of time (6 to 7 seconds, the most heavy ones get to about 10 seconds).

PS: As always the numbers are not very precise, I didn't waste too much time bench-marking this; though getting 30 seconds instead of 13 minutes multiple times by chance is extremely unlikely.